made a component. changed name of stopwatch hook file
