+++ src/components/app/app.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { Layout, Display, JustifyContent } from "ui";
-import { useStopwatch } from "hooks/stopwatch";
+import Stopwatch from "components/stopwatch";
 import "./app.scss";
 
 type AppProps = {
@@ -8,17 +8,6 @@ type AppProps = {
 };
 
 const App: React.FunctionComponent<AppProps> = (props: AppProps) => {
-  const { elapsedTime, start, stop, isRunning } = useStopwatch();
-  const buttonLabel = React.useMemo(() => (isRunning ? "Stop" : "Start"), [
-    isRunning,
-  ]);
-  const clickHandler = React.useCallback(() => (isRunning ? stop() : start()), [
-    start,
-    stop,
-    isRunning,
-  ]);
-  const elapsedSeconds = (elapsedTime / 1000).toFixed(2);
-
   return (
     <Layout
       display={Display.Flex}
@@ -31,8 +20,7 @@ const App: React.FunctionComponent<AppProps> = (props: AppProps) => {
         className="header"
       >
         <h1 className="shifting-rainbow-text">{props.message}</h1>
-        <h2 className="shifting-rainbow-text">{elapsedSeconds}</h2>
-        <button onClick={clickHandler}>{buttonLabel}</button>
+        <Stopwatch />
       </Layout>
     </Layout>
   );
