+++ src/components/stopwatch/index.ts
@@ -1,0 +1,3 @@
+import Stopwatch from "./stopwatch";
+
+export default Stopwatch;
