+++ src/components/stopwatch/stopwatch.tsx
@@ -1,0 +1,25 @@
+import React, { FunctionComponent } from "react";
+import { Layout } from "ui";
+import useStopwatch from "hooks/use-stopwatch";
+
+const Stopwatch: FunctionComponent = () => {
+  const { elapsedTime, start, stop, isRunning } = useStopwatch();
+  const buttonLabel = React.useMemo(() => (isRunning ? "Stop" : "Start"), [
+    isRunning,
+  ]);
+  const clickHandler = React.useCallback(() => (isRunning ? stop() : start()), [
+    start,
+    stop,
+    isRunning,
+  ]);
+  const elapsedSeconds = (elapsedTime / 1000).toFixed(2);
+
+  return (
+    <Layout>
+      <h2 className="shifting-rainbow-text">{elapsedSeconds}</h2>
+      <button onClick={clickHandler}>{buttonLabel}</button>
+    </Layout>
+  );
+};
+
+export default Stopwatch;
